extracted user&team stores to own autoconfig
